app/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
pydantic-settings
psycopg2-binary

# Required optional dependency for pydantic email validation
email-validator

# Opcionales según uso en el repositorio
alembic
passlib
argon2-cffi
python-jose[cryptography]
bcrypt
cryptography
requests